Refactor update_loo_subsample function; extract helper functions.
